Moved FreeRTOS Kernel to Submodule
Summary:
We now pull in the FreeRTOS kernel as a submodule. There are several
files that are not part of the kernel or are altered. To fix this there
are now patches for all files. These patches are applied by a script in
the patch folder. This script is called when sourcing the setenv.sh
script while building.

update-submodule: src/os/FreeRTOS-Kernel

Reviewed By: benwei13

Differential Revision: D27863908

fbshipit-source-id: f69978f