Move PR Template so it Doesn't Clutter PR Box
Summary:

When opening a PR from a link with a commit which has a matching commit
message, It fills the box with the templae and the commit message. Then
you have to go remove the template. This is very annoying. Move the
templates so they are still in the repo but don't generate the noise.

Test Plan:

Make a PR.

Reviewers:

Subscribers:

Tasks:

Tags: