Add clang-format Using Zephyrs Format File (#151)
Summary:
Add clang format to the CI so that it runs on PRs. Generally, if a file isn't properly formatted using clang-format it REALLY isn't formatted and generates a ton of errors.

Due to this I'm not going to annotate the PR and instead just leave a comment listing the offending file names.

Pull Request resolved: https://github.com/facebook/OpenBIC/pull/151

Test Plan:
Test on Opensource Repo to verify it functions correctly.

Comment when there are clang-format errors
{F698304107}

This is what the GitHub Action test results look like.
{F698304164}

Reviewed By: garnermic

Differential Revision: D33994432

Pulled By: GoldenBug

fbshipit-source-id: e03a7c7