Move PR Template so it Doesn't Clutter PR Box (#153)
Summary:
When opening a PR from a link with a commit which has a matching commit
message, It fills the box with the templae and the commit message. Then
you have to go remove the template. This is very annoying. Move the
templates so they are still in the repo but don't generate the noise.

Pull Request resolved: https://github.com/facebook/OpenBIC/pull/153

Test Plan: Make a PR.

Reviewed By: zhdaniel12

Differential Revision: D33994974

Pulled By: GoldenBug

fbshipit-source-id: e8c37a4