Remove Clang-Format Comment Generation Step
Summary:

When pull requests are made from a fork of a repo GitHub Tokens have
only read permissions by default. Unfortunately this means that a Github
action can't post a comment to the PR, like in the case of the
clang-format action.

Because of that I'm removing the code that generates the comment and
instead simply having the job for clang-formatting fail and requiring
the submitter go and investigate the cause.

Test Plan:

Run the action on the public Repo with and without files failing
clang-format checks.

Reviewers:

Subscribers:

Tasks:

Tags: CIT