Extend SDR_Full_sensor ID_str Array 16->18
Summary:
Several of the SDR_Full_sensor structures had an ID_str that was too
long (17 characters with a max length of 16).

Add 2 extra characters for a max size of 18 so we can fit the IDs
better.

```
meta-facebook/yv35-bb/src/platform/plat_sdr_table.c:497:3: warning: initializer-string for array of 'unsigned char' is too long
  497 |   "P1V2_BIC_STBY Vol",
```

Reviewed By: garnermic

Differential Revision: D35941622

fbshipit-source-id: 32306b53596796863ff185ead9c9744fbcb65bcc