Adding Werror to Builds so No Warnings Will be Introduced.
Summary:
Adding the Werror flag to Yv3.5 CL and BB platforms so that OpenBIC will
no longer build if there are errors introduced. This will make warnings
and errors more apparent and will stop them from being accidentally
introduced.

This will not be effected by warnings in the Zephyr Kernel because we
don't have much control over that.

Reviewed By: doranand

Differential Revision: D36262234

fbshipit-source-id: c036c79cc3c7eb82aa3dfe5dee974caef8d373d1