Replace All Internal Usage of Slave with Target
Summary:
Following company guidelines we should transition from using the slave
terminology to appropriate alternatives. I've gone through and replaced
all internal usage of slave with target instead, which seems to be the
consensus for i2c naming.

Unfortunately some of the slave terminology usage is in Zephyr itself so
it can't be changed when interacting/calling kernel functions.

Reviewed By: garnermic

Differential Revision: D36260104

fbshipit-source-id: b89749fd866f757429b5a8a5fa428cd859fd54e6