fby3.5: cl: Switch pages before querying PMBUS_IC_DEVICE_ID
Summary:
We're querying the voltage regulators for their IC_DEVICE_ID, but we're
not selecting the page of the PMBUS device beforehand. This can
manifest in weird behavior, but ultimately it would be a "Unknown VR
type".

Test Plan:
In QEMU I verified that with this fix, and a fix in QEMU's pmbus, that
the VR ID is read successfully for every VR.

Reviewers: goldenbug garnermic

Subscribers:

Tasks:

Tags: