Add clang-format Git Hook and Instructions
Summary:
Add a git hook that runs clang-format on a commit as well as
instructions on how to add the git hook.

This git hook will automatically stop a git commit and format the files that have improper formatting.
It is up to the user to review and add the format changes before attempting to commit the files again.

Reviewed By: garnermic

Differential Revision: D37802347

fbshipit-source-id: efba89cfa3d1ba9a432495c2b9eff5887627da73