Refactor Common plat_class Functions for Expansion Boards
Summary:
The expansion cards shared some code that should be extracted out into
common libraries.

In order to facilitate building the expansion board library and not
including extra logic for boards that don't use it; The library
directory was expanded in the CMakeLists.txt files so we can choose to
not include specific libraries.

In general globbing in cmake isn't great...

Reviewed By: garnermic

Differential Revision: D38228112

fbshipit-source-id: 1364addeffa06eb4a0487b9b74c3c5490eb7be77