Refactor Shell Source Code Structure
Summary:
This commit refactors the shell source code so that each sub command is
contained within it's own source file and header.

This makes the shell structure more expandable and limits the shell
files from growing too large.

New shell commands should be added in a similar way in the future.

Reviewed By: garnermic

Differential Revision: D38595995

fbshipit-source-id: ac9c2d14f534baa2f8d3915010bc4d0829392c51