rf: Set DC on delayed flag if the power already on
This patch fixes the VR sensors keeping not accessible after the RF FW
updating. It is because the vr_access function always checks the
is_DC_on_delayed flag, but this flag is only set in ISR_DC_STATE, which
can't be triggered if the power stage is already the last.

Tested:
After the BIC fw updating, the BIC shell command
`platform sensor list_all` shows the vr sensors are good.

Signed-off-by: Scron Chang <Scron.Chang@quantatw.com>