fby3.5: rf: Fixed CXL Update Issue
The last version of OpenBmc did not support the sector_end flag, so previous
bic firmware only assumed the fixed size per file.

This submission will use sector_end to determine the end sector of the image
for ensuring transmission successfully.

Test result:
The host could connect the CXL controller directly, and it does not need to do
AC_cycle.

Signed-off-by: Irene <Irene.Lin@quantatw.com>