common: I3C: Disable static addressing if it is not set in DTS
Summary:
Skip setting the static address if the "assigned-address" property is not assigned in slave mode.
This patch is intended to prevent a valid static address from affecting the behavior of the i3c slave.
The static address should only be enabled when the master wants to assign a dynamic address using setaasa or setdasa, or when the slave wants to operate in I2C mode.

Test Plan:
1. Build and test pass on fby35 system.

2. I3C communication works normally.
root@bmc-oob:~# bic-util slot2 0x18 0x1
00 80 31 04 02 BF 15 A0 00 00 00 00 00 00 00