common: ipmi: fix the mutex lock of ipmb read was not released (#1238)
Summary:
# Description
In our IPMI handler processing, we create a thread to handle operations and set a timeout of 3 seconds to prevent resource occupation. During this process, we may use IPMB read, however the timeout for waiting the IPMB response is also set to 3 seconds. This situation may result in the IPMI handler being kill due to exceeding the timeout, while the mutex lock for IPMB read has not been released.

# Motivation
To fix the mutex lock of ipmb read may not be released in some cases.

Pull Request resolved: https://github.com/facebook/OpenBIC/pull/1238

Test Plan: 1. Build and test pass on YV3.5 system. pass

Reviewed By: jagpalgill, tao-ren

Differential Revision: D47465811

fbshipit-source-id: 2a4ccbb75a1654951439e14ea55c62087bb177f3