fby3.5: gl: Fix BIC crash when dc off (#1443)
Summary:
# Description:
This issue was caused by the incorrect initialization of workqueue in interrupt function. The work function was initialized twice, if the interrupt received twice. And if the second function is initialed before the first function was done. the BIC will crash.

# Motivation:
To fix BIC crash when dc off.

Pull Request resolved: https://github.com/facebook/OpenBIC/pull/1443

Test Plan:
1. Build and test pass on OLP2.0 system. pass

2. Check whether BIC will crash when DC off pass

3. Stress power cycle about 500 times. pass

Reviewed By: jagpalgill

Differential Revision: D50826547

fbshipit-source-id: 0db83578d6b802a5876e2021d80784b9ed653c88