common: support to select the interface to communicate with BMC
Description:
Support to select I2C or I3C interface to communicate with BMC when BIC
is initializing.

Motivation:
We want to change the interface to communicate with BMC from I2C to I3C
in Yv4 project, but we think that the I3C function is not stable.
Therefore, we would like to support this feature so that we could select
the interface in BMC by setting SD CPLD IO expander and reboot BIC to
re-initialize, then we can test the I3C function.

Test logs:
Built pass for all project.