common: fix potential memory leak risk for kcs (#1975)
Summary:
# Description:
Use a temporary pointer for memory reallocation, then assign it to the original pointer.

# Motivation:
Without the temporary pointer, the original pointer could be lost if the reallocation fails, making it impossible to free the memory and potentially causing a memory leak.

Pull Request resolved: https://github.com/facebook/OpenBIC/pull/1975

Test Plan:
Build code - Pass
Tested on yv4 system for kcs-related functionality- Pass

Reviewed By: williamspatrick

Differential Revision: D62500187

fbshipit-source-id: 19c820c6ed2a89ec823c3bfe7ab54260f7221569