yv4-sd: custom pldm msg timeout and retry (#2520)
Summary:
# Description
Relate with the commit "common: change pldm msg timeout and retry". Custom the timeout and max retry to enhance the frequency of sending pldm msg.

# Motivation
Currently, the pldm msg timeout is for BIC FW update. But the timeout of msg is too long so judge pldm event type to use different timeout and max retry times.

Pull Request resolved: https://github.com/facebook/OpenBIC/pull/2520

Test Plan:
1. Build and test on YV4 system. pass
2. Validate by our QE team with 30 times blade AC cycle and 30 times blade AC off-on cycles.

Reviewed By: ifel, waffle2k

Differential Revision: D81427919

fbshipit-source-id: d35a2633d512fdd0244bd69d2f3d185fd67980fa