common: Add timestamp in oem pldm event (#2605)
Summary:
# Description
This PR has a dependency on https://gerrit.openbmc.org/c/openbmc/pldm/+/82142
New a timestamp in the pldm_addsel_data struct, so that BMC can use it to determine whether a SEL is duplicate.

# Motivation
Currently, BIC relies on receiving a response from BMC to determine whether an event is successful.
However, if BMC responds after the timeout, BIC treats the previous event as failed and retries.
Therefore, BMC will use the timestamp to distinguish whether it receives a retried event or a new event to avoid recording same SEL again.

Pull Request resolved: https://github.com/facebook/OpenBIC/pull/2605

Test Plan:
1. Build and test on YV4 system. pass
2. Validate by our QE team with 51 times of BIOS update, 52 times of DC cycle and 45 times ACPI reset.

Reviewed By: wangx6f

Differential Revision: D87195951

fbshipit-source-id: dbca0a39eb7ab552ad68075df29c67bf73da342c